changed redux for scenario creation
