+++ client/src/containers/scenarios/edit/component.tsx
@@ -1,26 +1,23 @@
 import { useCallback } from 'react';
-import { useDebounceCallback } from '@react-hook/debounce';
-import { useForm } from 'react-hook-form';
-import cx from 'classnames';
-import { useRouter } from 'next/router';
-import toast from 'react-hot-toast';
-import * as yup from 'yup';
-import { yupResolver } from '@hookform/resolvers/yup';
 
-import { useScenario, useUpdateScenario } from 'hooks/scenarios';
-import { useInterventions } from 'hooks/interventions';
+import cx from 'classnames';
+import { useQuery } from 'react-query';
+import { apiService } from 'services/api';
+import { useRouter } from 'next/router';
 
 import { useAppDispatch, useAppSelector } from 'store/hooks';
+
 import { setSubContentCollapsed } from 'store/features/analysis/ui';
-import { setScenarioTab, scenarios } from 'store/features/analysis/scenarios';
+import { scenarios, setScenarioTab } from 'store/features/analysis/scenarios';
+
+import { useInterventions } from 'hooks/interventions';
 
 import Button from 'components/button';
 import InterventionsList from 'containers/interventions/list';
-import { Label, Input, Textarea } from 'components/forms';
+import Label from 'components/forms/label';
+import Textarea from 'components/forms/textarea';
 import GrowthList from 'containers/growth/list/component';
 import { PlusIcon } from '@heroicons/react/solid';
-
-import type { ErrorResponse } from 'types';
 
 const items = [
   {
@@ -37,71 +34,28 @@ const items = [
   },
 ];
 
-const schemaValidation = yup.object({
-  title: yup.string().min(2).required(),
-  description: yup.string(),
-});
-
 const ScenariosNewContainer: React.FC = () => {
-  const { query } = useRouter();
-
-  const dispatch = useAppDispatch();
-  const { scenarioCurrentTab, currentScenario } = useAppSelector(scenarios);
-
-  const {
-    register,
-    getValues,
-    formState: { isValid, errors },
-  } = useForm({
-    mode: 'onChange',
-    resolver: yupResolver(schemaValidation),
-  });
-
-  const { data: scenario, isLoading } = useScenario(currentScenario as string, { sort: 'title' });
   const { data: interventions } = useInterventions({ sort: query.sortBy as string });
-
-  const updateScenario = useUpdateScenario();
-
+  const dispatch = useAppDispatch();
   const handleNewScenarioFeature = useCallback(() => {
     dispatch(setSubContentCollapsed(false));
   }, [dispatch]);
 
-  const handleTab = useCallback((step) => dispatch(setScenarioTab(step)), [dispatch]);
+  const { scenarioCurrentTab } = useAppSelector(scenarios);
 
-  const handleChange = useDebounceCallback(
-    useCallback(() => {
-      const { id } = scenario;
-      if (isValid) {
-        updateScenario.mutate(
-          { id, data: getValues() },
-          {
-            onSuccess: () => {
-              toast.success('Your changes were successfully saved.');
-            },
-            onError: (error: ErrorResponse) => {
-              const { errors } = error.response?.data;
-              errors.forEach(({ title }) => toast.error(title));
-            },
-          },
-        );
-      }
-    }, [scenario, isValid, updateScenario, getValues]),
-    500,
-  );
+  const handleTab = useCallback((step) => dispatch(setScenarioTab(step)), [dispatch]);
 
   return (
     <>
-      <form className="z-20">
-        <Input
+      <form action="#" method="POST" className="z-20">
+        <input
           type="text"
-          {...register('title')}
+          name="title"
+          id="title"
           placeholder="Untitled"
-          defaultValue={scenario.title}
           aria-label="Scenario title"
-          className="flex-1 block w-full md:text-2xl sm:text-sm border-none text-gray-900 p-0 mb-6"
-          onInput={handleChange}
-          disabled={isLoading}
-          error={errors.title?.message}
+          autoFocus
+          className="flex-1 block w-full md:text-2xl sm:text-sm border-none text-gray-400 p-0 font-semibold mb-6"
         />
 
         <div className="sm:col-span-6">
@@ -109,12 +63,11 @@ const ScenariosNewContainer: React.FC = () => {
             Scenario description <span className="text-gray-500">(optional)</span>
           </Label>
           <Textarea
-            {...register('description')}
+            id="description"
+            name="description"
             rows={3}
             className="w-full"
-            defaultValue={scenario.description}
-            disabled={isLoading}
-            error={errors.description?.message}
+            defaultValue=""
           />
         </div>
       </form>
@@ -126,7 +79,7 @@ const ScenariosNewContainer: React.FC = () => {
                 <button
                   type="button"
                   className={cx({
-                    'border-b-2 border-green-700': scenarioCurrentTab === 'interventions',
+                    'border-b-2 border-green-700': scenarioCurrentTab == 'interventions',
                   })}
                   onClick={() => handleTab('interventions')}
                 >
@@ -135,7 +88,7 @@ const ScenariosNewContainer: React.FC = () => {
 
                 <button
                   type="button"
-                  className={cx({ 'border-b-2 border-green-700': scenarioCurrentTab === 'growth' })}
+                  className={cx({ 'border-b-2 border-green-700': scenarioCurrentTab == 'growth' })}
                   onClick={() => handleTab('growth')}
                 >
                   Growth rates ({items.length})
@@ -147,8 +100,8 @@ const ScenariosNewContainer: React.FC = () => {
               </Button>
             </div>
           </div>
-          {scenarioCurrentTab === 'growth' && <InterventionsList items={interventions} />}
-          {scenarioCurrentTab === 'interventions' && <GrowthList items={items} />}
+          {scenarioCurrentTab == 'interventions' && <InterventionsList items={interventions} />}
+          {scenarioCurrentTab == 'growth' && <GrowthList items={items} />}
         </div>
       </div>
     </>
